pyproject.toml
[project]
name = "phuong-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiokafka>=0.12.0",
    "fastapi[standard]>=0.115.8",
    "gradio>=5.34.2",
    "gradio-webrtc>=0.0.31",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pika>=1.3.2",
    "pytest>=8.3.4",
    "rabbitmq>=0.2.0",
    "redis>=6.2.0",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.13.4",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    "htmlmin>=0.1.12",
    "lxml-html-clean>=0.4.2",
    "newspaper3k>=0.2.8",
    "simplejson>=3.20.1",
    "qdrant-client>=1.15.1",
    "openai>=2.2.0",
    "sqlalchemy>=2.0.45",
    "psycopg2>=2.9.11",
    "python-jose>=3.5.0",
    "schedule>=1.2.2",
    "transformers>=4.57.3",
    "passlib>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.ruff.mccabe]
max-complexity = 10
